Simplify assembly macro to only output softcore operations
As a library this crate should not be too opinionated about what code it
generates, especially it should not impose features on the user crates.
It is easy for a user crate to create its own wrapping macro that emits
both the assembly and the softcore equivalent if required. Thus in the
future we leave this choice to the user.