Test one of Miralis' trap handlers
This commit adds one of Miralis' trap handler as a test. Turns out the
handler had a bug! This version pushed contain the fix.

The bug was that a register was temporarily saved in mepc, however bit 0
of mepc is hardwired to 0, so we were losing one bit. This is a good
proof of concept of the value of softcore-asm :)