Use a &raw const reference to symbols
Unfurtunately it is hard to guess if a symbol address will be used to
mutate data or not in an assembly snippet. The safe solution would be to
require a mut reference, however this prevent the use of read-only
values (such as non-mut statics).
For now we will use a &raw const instead, but the assembly author should
ensure that the symbol is exposed as mutable when performing mutation
from the assembly, this won't be enforced by the compiler!

We will probably need to come back and revisit this decision later.