Add initial assembly parser skeleton
To correctly parse assembly we need a more powerful way to parse that
relying on regexes. For instance, loads and stores can include arbitrary
arithmetic expressions as part of the offset, and therefore the parser
must understand operator precedence.

This commit adds the initial scaffolding of a proper parser, based on
the Pest parser generator. Currently, the parser only support numerical
expressions and is not yet integrated as part of the macro.