Add softcore argument to asm macro
The new argument is required to indicate which function to call in order
to get access to the core. Typically something of the form
`SOFT_CORE.with_borrow_mut`.

It is expected that users will define a macro that passes the argument
automatically for better ergonomics. Here is an example of such a macro:

```rs
macro_rules! rasm {
    ($($asm:tt)*) => {
        softcore_asm::rasm!(
            $($asm)*,
            softcore(SOFT_CORE.with_borrow_mut)
        )
    };
}
```