Add initial support for calling Rust function from assembly
Add support for the "call" pseudo instruction to call a Rust function
symbol. We define a new trait, `AsmCallable` for `extern "C" f()` with
0, 1, or 2 arguments (can be extended as needed in the future).
Unfortunately, Rust can not yet automatically coerce a function into the
function type with the appropriate arity (i.e. number of arguments), so
we will need explicit user annotations to make the conversion explicit
in the macro. For now, we only support functions with 0 arguments.