Add initial support for attribute
This commit adds support for parsing assembly attributes, in the shape
of `// #[attribute_name(<arg1>, <arg2>, ...)]`. Attributes apply to the
next following instruction, and can be used in the future to let the
user inform the macro about how to generate Rust code, such as adding
type annotations when necessary.

No attribute has been implemented yet, only parsing so far.