Add support for parsing `#[abi("C", N)]` attribute
We can now use this attribute to specify the proper abi and number of
parameters for Rust function calls from assembly, which is required
because the current Rust type system can not coalesce a function type
`extern "C" fn(...) -> ... {my_func}` into the more generic
`extern "C" fn(...) -> ...` type. This attribute lets the user
explicitly specify those argument, which enables the macro to do a
cast with the appropriate arity (number of arguments).