Add support for nested assembly calls
With this commit, an assembly snippet can call a Rust function that uses
another assembly snippet internally.
To support that feature this commit changes the code generation to
release the Mutex/RefCell before calling into a Rust function. The
Mutex/RefCell must also be release on control flow, because an `if`
might no know which branches might perform a call.

The current code generation is not optimal in terms of Mutex/RefCell
acquire/releases. Future work can improve on it if necessary.