Add support for #[replace_with("instr")] attribute
This attribute can be used to replace an assembly instruction by
another. It is useful when using a macro to emit either softcore-asm
assembly or real Rust assembly, as Rust will ignore the attribute. This
allows emitting a different instruction in the two case, which enables
skipping or replacing instructions for tests.