Inline CSR ID into generated Rust code
Prior to this commit the CSRs ID were looked-up by their name using one
of the RISC-V model functions (`csr_name_map_backwards`). This causes
overhead at runtime, and was also causing trouble to Kani, a Rust model
checker we use to verify system software.
This commit instead performs the lookup in the proc-macro, and directly
output the integer as part of the generated Rust code.

There is a caveat, however. The Sail-to-Rust compiler only emits 'static
lifetime for strings at the current time, due to the lack of a proper
analysis pass. As a result, we need to make the string's lifetime static
in the macro, which we do by leaking the string. This is of course bad,
and we should fix it, but proc-macro are expected to be short lived, so
hopefully the consequences will be manageable until we fix the root
cause of that issue.